requirements.txt
# Divestment SOW Auditor - Requirements
# Install: pip install -r requirements.txt

# Core Dependencies
customtkinter==5.2.2
python-dotenv==1.0.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.16.3

# DOCX Processing  
python-docx==1.1.0

# OCR (Optional but recommended)
pytesseract==0.3.10
Pillow==10.2.0

# API Requests
requests==2.31.0

# PDF Report Generation
reportlab==4.0.9

# Email
# (Built-in: smtplib, email)

# Note: For OCR functionality, you also need to install Tesseract OCR:
# Windows: https://github.com/UB-Mannheim/tesseract/wiki
# Mac: brew install tesseract
# Linux: sudo apt-get install tesseract-ocr
